Not entirely sure what happened after I updated my project after the UserController branch merged but everything was duplicated after that so I just deleted my old copies and moved the updated ones in. Now it thinks I made changes to the controllers. Anyways, navigation bar implemented on the pages after login, added some css because I was bored.
