--- CLC-Order-Management-Application/WebContent/META-INF/MANIFEST.MF
@@ -1,5 +1,0 @@
-Manifest-Version: 1.0
-Created-By: IntelliJ IDEA
-Built-By: JCH
-Build-Jdk: 11.0.6
-
