--- CLC-Order-Management-Application/WebContent/WEB-INF/dispatcher-servlet.xml
@@ -1,45 +1,0 @@
-<beans xmlns="http://www.springframework.org/schema/beans"
-	xmlns:context="http://www.springframework.org/schema/context"
-	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:mvc="http://www.springframework.org/schema/mvc"
-	xsi:schemaLocation="
-http://www.springframework.org/schema/beans
-http://www.springframework.org/schema/beans/spring-beans-3.0.xsd
-http://www.springframework.org/schema/context
-http://www.springframework.org/schema/context/spring-context-3.0.xsd
-http://www.springframework.org/schema/mvc
-http://www.springframework.org/schema/mvc/spring-mvc.xsd">
- 
-	<context:component-scan base-package="com.gcu" />
-
-	<mvc:annotation-driven/>
-	<bean
-		class="org.springframework.web.servlet.view.InternalResourceViewResolver">
-		<property name="prefix">
-			<value>/WEB-INF/views/</value>
-		</property>
-		<property name="suffix">
-			<value>.jsp</value>
-		</property>
-	</bean>
-	
-	<bean id = "userValidator" class = "com.gcu.validators.UserValidator"/>
-	<bean id = "orderValidator" class = "com.gcu.validators.OrderValidator"/>
-
-	<bean id="ds" class="org.springframework.jdbc.datasource.DriverManagerDataSource">
-		<property name="driverClassName" value="com.mysql.jdbc.Driver"/>
-		<property name="url" value="jdbc:mysql://clc-order.commclm1m94c.us-east-2.rds.amazonaws.com:3306/CST323"/>
-		<property name="username" value="admin"/>
-		<property name="password" value="adminpassword"/>
-	</bean>
-
-	<bean id="jt" class="org.springframework.jdbc.core.JdbcTemplate">
-		<property name="dataSource" ref="ds"/>
-	</bean>
-
-	<bean id="user_dao" class="com.gcu.dao.UserDao">
-		<property name="template" ref="jt"/>
-	</bean>
-	<bean id="order_dao" class="com.gcu.dao.OrderDao">
-		<property name="template" ref="jt"/>
-	</bean>
-</beans>
