--- CLC-Order-Management-Application/WebContent/WEB-INF/views/helloworld.jsp
@@ -1,20 +1,0 @@
-<%@ page language="java" contentType="text/html; charset=ISO-8859-1" pageEncoding="ISO-8859-1"%>
-<!DOCTYPE html PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN" "http://www.w3.org/TR/html4/loose.dtd">
-<html>
-<head>
-<meta http-equiv="Content-Type" content="text/html; charset=ISO-8859-1">
-<title>Welcome ${user.firstName}!</title>
-</head>
-<body>
-	<center>
-		<h2>
-			${message} ${user.firstName}!
-		</h2>
-		<div>
-			<a href="placeOrder">Place Order</a>
-			<a href="viewOrders">View Orders</a>
-		</div>
-	</center>
-
-</body>
-</html>
