--- CLC-Order-Management-Application/WebContent/index.jsp
@@ -1,19 +1,0 @@
-<%@ page language="java" contentType="text/html; charset=ISO-8859-1"
- 
-pageEncoding="ISO-8859-1"%>
-<!DOCTYPE html PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN" "http://www.w3.org/TR/html4/loose.dtd">
-<html>
-<head>
-<meta http-equiv="Content-Type" content="text/html; charset=ISO-8859-1">
-<title>CST-323 Order Management Application</title>
-</head>
-<body>
- 
-	<center>
-		<h2>Welcome to the Order Management Application Registration!</h2>
-		<h3>
-			<a href = "register">Click Here to Register An Account!</a>
-		</h3>
-	</center>
-</body>
-</html>
