--- CLC-Order-Management-Application/src/Model/Login.java
@@ -1,27 +1,0 @@
-package Model;
-
-public class Login 
-{
-	private String username;
-	private String password;
-	
-	public String getUsername()
-	{
-		return username;
-	}
-	
-	public String getPassword()
-	{
-		return password;
-	}
-	
-	public void setUsername(String incUser)
-	{
-		this.username = incUser;
-	}
-	
-	public void setPassword(String incPass)
-	{
-		this.password = incPass;
-	}
-}
