+++ CLC-Order-Management-Application/src/main/webapp/WEB-INF/dispatcher-servlet.xml
@@ -1,6 +1,7 @@
 <beans xmlns="http://www.springframework.org/schema/beans"
 	xmlns:context="http://www.springframework.org/schema/context"
-	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:mvc="http://www.springframework.org/schema/mvc"
+	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+	xmlns:mvc="http://www.springframework.org/schema/mvc"
 	xsi:schemaLocation="
 http://www.springframework.org/schema/beans
 http://www.springframework.org/schema/beans/spring-beans-3.0.xsd
@@ -10,7 +11,7 @@ http://www.springframework.org/schema/mvc
 http://www.springframework.org/schema/mvc/spring-mvc.xsd">
  
 	<context:component-scan base-package="com.gcu" />
-
+	<mvc:resources mapping="/resources/**" location="/resources/"/>
 	<mvc:annotation-driven/>
 	<bean
 		class="org.springframework.web.servlet.view.InternalResourceViewResolver">
