+++ CLC-Order-Management-Application/src/main/webapp/WEB-INF/views/login.jsp
@@ -1,25 +1,17 @@
 <%@taglib uri="http://www.springframework.org/tags/form" prefix="form"%>
 <%@ page language="java" contentType="text/html; charset=ISO-8859-1"
     pageEncoding="ISO-8859-1"%>
+<%@ taglib uri="http://java.sun.com/jsp/jstl/core" prefix="c"%>
 <!DOCTYPE html PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN" "http://www.w3.org/TR/html4/loose.dtd">
 <html>
-<style>
-	.error{
-		color: #ff0000;
-	}
-	.errorblock{
-		color: #000;
-		background-color: #ffEEEE;
-		border: 3px solid #ff0000;
-		padding: 8px;
-		margin: 16px;
-	}
-</style>
+
 <head>
+<link type="text/css" rel="stylesheet" href="<c:url value="/resources/css/styles.css"/>"/>
 <meta http-equiv="Content-Type" content="text/html; charset=ISO-8859-1">
 <title>Login</title>
 </head>
 <body>
+<div>
 Please Login to Continue:
 <form:form method="POST" action = "displayHome" modelAttribute="user">
 <form:errors path ="*" cssClass = "errorblock" element = "div"/>
@@ -40,5 +32,6 @@ Please Login to Continue:
         </tr>
     </table>
 </form:form>
+</div>
 </body>
 </html>
