+++ CLC-Order-Management-Application/src/main/webapp/WEB-INF/views/register.jsp
@@ -1,27 +1,16 @@
 <%@taglib uri="http://www.springframework.org/tags/form" prefix="form"%>
 <%@ page language="java" contentType="text/html; charset=ISO-8859-1"
     pageEncoding="ISO-8859-1"%>
+<%@ taglib uri="http://java.sun.com/jsp/jstl/core" prefix="c"%>
 <!DOCTYPE html PUBLIC "-//W3C//DTD HTML 4.01 Transitional//EN" "http://www.w3.org/TR/html4/loose.dtd">
 <html>
-
-<style>
-	.error{
-		color: #ff0000;
-	}
-	.errorblock{
-		color: #000;
-		background-color: #ffEEEE;
-		border: 3px solid #ff0000;
-		padding: 8px;
-		margin: 16px;
-	}
-</style>
-
 <head>
-<meta http-equiv="Content-Type" content="text/html; charset=ISO-8859-1">
+    <link type="text/css" rel="stylesheet" href="<c:url value="/resources/css/styles.css"/>"/>
+    <meta http-equiv="Content-Type" content="text/html; charset=ISO-8859-1">
 <title>Register</title>
 </head>
 <body>
+<div>
 Please Register to Continue:
 <form:form method="POST" action = "helloworld" modelAttribute="user">
 <form:errors path ="*" cssClass = "errorblock" element = "div"/>
@@ -62,5 +51,6 @@ Please Register to Continue:
         </tr>
     </table>
 </form:form>
+</div>
 </body>
 </html>
