+++ CLC-Order-Management-Application/src/main/webapp/resources/css/styles.css
@@ -1,0 +1,44 @@
+body
+{
+    margin: 0;
+}
+.error{
+    color: #ff0000;
+}
+.errorblock{
+    color: #000;
+    background-color: #ffEEEE;
+    border: 3px solid #ff0000;
+    padding: 8px;
+    margin: 16px;
+}
+.nav
+{
+    list-style-type: none;
+    margin: 0;
+    padding: 0;
+    overflow: hidden;
+    background-color: #333;
+}
+li {
+    float: left;
+}
+li a {
+    display: block;
+    color: white;
+    text-align: center;
+    padding: 14px 16px;
+    text-decoration: none;
+}
+
+li a:hover:not(.active) {
+    background-color: #20a8ff;
+}
+
+.active {
+    background-color: #1d7abf;
+}
+div
+{
+    margin: 5px;
+}
